Move config to src
